improv: Subclass GtkApplicationWindow
